fix ao require wifi ou dados moveis. se a versão for diferente atualzia o studie

+++ android/movesense-app/app/src/main/java/com/umaia/movesense/data/responses/studies_response/StudyVersionResponse.kt
@@ -1,5 +1,5 @@
 package com.umaia.movesense.data.responses.studies_response
 
 data class StudyVersionResponse(
-    val version: Int
+    val version: Double
 )
